Add /health-check
We have a HealthCheck object that does a bunch of checks and information gathering.  We also have a rack middleware that exposes this information as JSON when we visit /health-check.

We tried using a simple controller to render the JSON but it didn't work in some of the failure cases.  For example one of our checks is that the DB is working and rails provides rack middleware that will break if the DB is not present (ActiveRecord::QueryCache) so our controller action would never be hit.  Using rack middleware we can inject our healthcheck where-ever we want in the stack to run the healthcheck before rails will break.